[arccore+arcane] Move and rename 'arcane/utils/Limits.h' to 'arccore/base/FloatInfo.h'.
