+++ arccore/src/base/arccore/base/ArccoreGlobal.cc
@@ -32,6 +32,7 @@
 #include "arccore/base/ParallelLoopOptions.h"
 #include "arccore/base/ISymbolizerService.h"
 #include "arccore/base/internal/IDynamicLibraryLoader.h"
+#include "arccore/base/FloatInfo.h"
 
 #include <iostream>
 #include <cstring>
